Fix <script> index, Update README
Fixed the index of script again, and added code to auto-correct itself.
Updated the README to reflect the new changes I made today/yesterday.

Up next is to fix the errors that are being raised. I might have to do some reading on best practices for this type of stuff in Python.